AlteraPlatformPkg: Fogbugz #314587: Update DDR Scheduler ReadLatency calculation
The ReadLatency value is supposed to be set to the latency
out to SDRAM and back. Current calculation is considering
only the SDRAM portion of the total latency, and therefore
it is very under valued.  This will result in the DDR Scheduler
issuing a few requests and then it will stall waiting for
these requests to return, therefore limiting the bandwidth.
Besides, the current ReadLatency value is in processor clock unit.
Hence, it must be divided by 2 for hard memory controller clock unit.

Contributed-under: TianoCore Contribution Agreement 1.0

Signed-off-by: VIET NGA DAO <vndao@altera.com>